[renderer-ff] decided not to feature flag pure markdown
