+++ packages/perseus/src/perseus-markdown.new.tsx
@@ -1,10 +1,7 @@
 // TODO(LEMS-4304): feature flag cleanup - rename this file to perseus-markdown.tsx.
 // WARNING!!!! THERE ARE ACTIVE CHANGES HAPPENING IN PERSEUS MARKDOWN!
 // This file is the new perseus markdown that will replace the old perseus markdown.
-import {
-    pureMarkdownRulesNew as pureMarkdownRules,
-    traverseContent,
-} from "@khanacademy/pure-markdown";
+import {pureMarkdownRules, traverseContent} from "@khanacademy/pure-markdown";
 import SimpleMarkdown from "@khanacademy/simple-markdown";
 import * as React from "react";
 import _ from "underscore";
